Added Admin panel, editing and delete users. Access only for admin

+++ blog/src/main/java/softuniBlog/controller/ArticleController.java
@@ -2,8 +2,10 @@ package softuniBlog.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.AnonymousAuthenticationFilter;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -48,6 +50,13 @@ public class ArticleController {
         if(!this.articleRepository.exists(id)){
             return "redirect:/";
         }
+
+        if(!(SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken)){
+            UserDetails principal = (UserDetails)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            User entityUser = this.userRepository.findByEmail(principal.getUsername());
+            model.addAttribute("user", entityUser);
+        }
+
         Article article = this.articleRepository.findOne(id);
 
         model.addAttribute("article", article);
@@ -63,6 +72,9 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+        if(!isUserAuhorOrAdmin(article)){
+            return "redirect:/article/"+id;
+        }
         model.addAttribute("view", "article/edit");
         model.addAttribute("article", article);
 
@@ -76,6 +88,9 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+        if(!isUserAuhorOrAdmin(article)){
+            return "redirect:/article/"+id;
+        }
         article.setContent(articleBindingModel.getContent());
         article.setTitle(articleBindingModel.getTitle());
 
@@ -90,7 +105,9 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
-
+        if(!isUserAuhorOrAdmin(article)){
+            return "redirect:/article/"+id;
+        }
         model.addAttribute("view", "article/delete");
         model.addAttribute("article", article);
 
@@ -104,8 +121,20 @@ public class ArticleController {
             return "redirect:/";
         }
         Article article = this.articleRepository.findOne(id);
+        if(!isUserAuhorOrAdmin(article)){
+            return "redirect:/article/"+id;
+        }
+
         this.articleRepository.delete(article);
 
         return "redirect:/";
     }
+
+    public boolean isUserAuhorOrAdmin(Article article){
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+        return userEntity.isAdmin() || userEntity.isAuthor(article);
+    }
+
+
 }
